feat: add hive control system
