+++ hive-generator-service/src/main/java/io/pockethive/Topology.java
