+++ hive-generator-service/src/main/java/io/pockethive/generator/Application.java
