+++ hive-generator-service/src/main/java/io/pockethive/generator/RabbitConfig.java
