+++ hive-generator-service/src/main/resources/application.yml
