+++ hive-moderator-service/src/main/java/io/pockethive/moderator/RabbitConfig.java
