+++ hive-postprocessor-service/src/main/java/io/pockethive/postprocessor/RabbitConfig.java
