+++ hive-processor-service/src/main/java/io/pockethive/processor/Application.java
