+++ hive-processor-service/src/main/java/io/pockethive/processor/RabbitConfig.java
